\name{logevityTools-package}
\alias{logevityTools-package}
\alias{logevityTools}
\docType{package}
\title{
\packageTitle{logevityTools}
}
\description{
\packageDescription{logevityTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{logevityTools}
\packageIndices{logevityTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{logevityTools}

Maintainer: \packageMaintainer{logevityTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
