\name{runLimma}
\alias{runLimma}
\alias{sampleList}
\title{
DEG analysis with limma in batch mode
}
\description{
The \code{sampleList} function extracts the sample comparisons (contrasts) 
from the CMAP annotation table and stores them as a list. The \code{runLimma}
function performs the actual DEG analysis with the \code{limma} package for
all comparisons defined by \code{sampleList}.
}
\usage{
sampleList(cmap, myby)

runLimma(df, comp_list, fdr = 0.05, foldchange = 1, verbose = TRUE, affyid=NULL)
}
\arguments{
  \item{df}{
\code{data.frame} containing normalized intensity values. Row names
are the probe set identifiers of the corresponding Affymetrix chip.
}
  \item{comp_list}{
Sample comparison \code{list} generated by \code{sampleList}.
}
  \item{fdr}{
FDR cutoff filter value on scale between 0 to 1.
}
  \item{foldchange}{
log2 fold change filter.
}
  \item{affyid}{
If \code{affyid=NULL}, a binary DEG matrix will be returned; if \code{affyid} is 
assigned a character vector of Affy IDs, a list will be returned containing for 
each treatment the \code{topTable} output subsetted by the corresponding Affy IDs. 
}
  \item{verbose}{
Function prints status output if \code{verbose=TRUE}, while \code{FALSE} suppresses it.
}
  \item{cmap}{
Input \code{data.frame} defining sample comparisons, such as CMAP annotation file
(cmap_instances_02.txt).
}
  \item{myby}{
If \code{myby=CMP} the comparisons will consider only the compound information, while \code{myby=CMP_CELL} also 
the cell types used in the assyas.
}
}
\value{
Function \code{sampleList} returns a \code{list} of \code{character} vectors.

If \code{affyid=NULL}, the function \code{runLimma} returns a list with the
following three components. \code{DEG}: binary matrix where DEGs meeting the
cutoff settings are indicated by 1s; \code{logFC}: matrix with corresponding
logFC values; \code{FDR} matrix with corresponding FDR values. If \code{affyid}
is assigned a character vector of Affy IDs, then the function returns a list
containing for each treatment the \code{topTable} output subsetted by the
corresponding Affy IDs.
}

\author{
Thomas Girke
}

\examples{

\dontrun{
## Generate CEL file list for treatment vs. control comparisons
cmap <- read.delim("./data/cmap_instances_02.txt", check.names=FALSE)
comp_list <- sampleList(cmap, myby="CMP") 
comp_list <- sampleList(cmap, myby="CMP_CELL") # includes cell type info

## Load normalized expression data
chiptype_dir <- unique(readRDS("./data/chiptype.rds"))
df1 <- readRDS(paste0("data/", chiptype_dir[1], "/", "all_mas5exprs.rds"))
df2 <- readRDS(paste0("data/", chiptype_dir[2], "/", "all_mas5exprs.rds"))
df3 <- readRDS(paste0("data/", chiptype_dir[3], "/", "all_mas5exprs.rds"))
affyid <- rownames(df1)[rownames(df1) %in% rownames(df2)]; affyid <- affyid[affyid %in% rownames(df3)]
df <- cbind(df1[affyid,], df2[affyid,], df3[affyid,])

## DEG analysis with Limma
degMA <- runLimma(df, comp_list, fdr=0.10, foldchange=1, verbose=TRUE) 

}
}
\keyword{ utilities }
